Build fixes.
