+++ Code/Engine/Foundation/IO/Implementation/Posix/MemoryMappedFile_posix.h
@@ -1,6 +1,7 @@
 #include <Foundation/FoundationPCH.h>
 EZ_FOUNDATION_INTERNAL_HEADER
 
+#include <Foundation/Strings/String.h>
 #include <Foundation/IO/MemoryMappedFile.h>
 #include <Foundation/Logging/Log.h>
 #include <Foundation/Strings/PathUtils.h>
