+++ Code/Engine/Foundation/Types/Implementation/Variant_inl.h
@@ -386,168 +386,6 @@ T ezVariant::ConvertTo(ezResult* out_pConversionStatus /* = nullptr*/) const
   return result;
 }
 
-// for some reason MSVC does not accept the template keyword here
-#if EZ_ENABLED(EZ_COMPILER_MSVC_PURE)
-#  define CALL_FUNCTOR(functor, type) return functor.operator()<type>(std::forward<Args>(args)...)
-#else
-#  define CALL_FUNCTOR(functor, type) return functor.template operator(std::forward<Args>(args)...)<type>()
-#endif
-
-template <typename Functor, class... Args>
-auto ezVariant::DispatchTo(Functor& functor, Type::Enum type, Args&&... args)
-{
-  switch (type)
-  {
-    case Type::Bool:
-      CALL_FUNCTOR(functor, bool);
-      break;
-
-    case Type::Int8:
-      CALL_FUNCTOR(functor, ezInt8);
-      break;
-
-    case Type::UInt8:
-      CALL_FUNCTOR(functor, ezUInt8);
-      break;
-
-    case Type::Int16:
-      CALL_FUNCTOR(functor, ezInt16);
-      break;
-
-    case Type::UInt16:
-      CALL_FUNCTOR(functor, ezUInt16);
-      break;
-
-    case Type::Int32:
-      CALL_FUNCTOR(functor, ezInt32);
-      break;
-
-    case Type::UInt32:
-      CALL_FUNCTOR(functor, ezUInt32);
-      break;
-
-    case Type::Int64:
-      CALL_FUNCTOR(functor, ezInt64);
-      break;
-
-    case Type::UInt64:
-      CALL_FUNCTOR(functor, ezUInt64);
-      break;
-
-    case Type::Float:
-      CALL_FUNCTOR(functor, float);
-      break;
-
-    case Type::Double:
-      CALL_FUNCTOR(functor, double);
-      break;
-
-    case Type::Color:
-      CALL_FUNCTOR(functor, ezColor);
-      break;
-
-    case Type::ColorGamma:
-      CALL_FUNCTOR(functor, ezColorGammaUB);
-      break;
-
-    case Type::Vector2:
-      CALL_FUNCTOR(functor, ezVec2);
-      break;
-
-    case Type::Vector3:
-      CALL_FUNCTOR(functor, ezVec3);
-      break;
-
-    case Type::Vector4:
-      CALL_FUNCTOR(functor, ezVec4);
-      break;
-
-    case Type::Vector2I:
-      CALL_FUNCTOR(functor, ezVec2I32);
-      break;
-
-    case Type::Vector3I:
-      CALL_FUNCTOR(functor, ezVec3I32);
-      break;
-
-    case Type::Vector4I:
-      CALL_FUNCTOR(functor, ezVec4I32);
-      break;
-
-    case Type::Vector2U:
-      CALL_FUNCTOR(functor, ezVec2U32);
-      break;
-
-    case Type::Vector3U:
-      CALL_FUNCTOR(functor, ezVec3U32);
-      break;
-
-    case Type::Vector4U:
-      CALL_FUNCTOR(functor, ezVec4U32);
-      break;
-
-    case Type::Quaternion:
-      CALL_FUNCTOR(functor, ezQuat);
-      break;
-
-    case Type::Matrix3:
-      CALL_FUNCTOR(functor, ezMat3);
-      break;
-
-    case Type::Matrix4:
-      CALL_FUNCTOR(functor, ezMat4);
-      break;
-
-    case Type::Transform:
-      CALL_FUNCTOR(functor, ezTransform);
-      break;
-
-    case Type::String:
-      CALL_FUNCTOR(functor, ezString);
-      break;
-
-    case Type::StringView:
-      CALL_FUNCTOR(functor, ezStringView);
-      break;
-
-    case Type::DataBuffer:
-      CALL_FUNCTOR(functor, ezDataBuffer);
-      break;
-
-    case Type::Time:
-      CALL_FUNCTOR(functor, ezTime);
-      break;
-
-    case Type::Uuid:
-      CALL_FUNCTOR(functor, ezUuid);
-      break;
-
-    case Type::Angle:
-      CALL_FUNCTOR(functor, ezAngle);
-      break;
-
-    case Type::VariantArray:
-      CALL_FUNCTOR(functor, ezVariantArray);
-      break;
-
-    case Type::VariantDictionary:
-      CALL_FUNCTOR(functor, ezVariantDictionary);
-      break;
-
-    case Type::TypedObject:
-      CALL_FUNCTOR(functor, ezTypedObject);
-      break;
-
-    default:
-      EZ_REPORT_FAILURE("Could not dispatch type '{0}'", type);
-      // Intended fall through to disable warning.
-    case Type::TypedPointer:
-      CALL_FUNCTOR(functor, ezTypedPointer);
-      break;
-  }
-}
-
-#undef CALL_FUNCTOR
 
 /// private methods
